requirements.txt
selenium~=4.8.2
pytest~=7.2.2
allure-pytest~=2.13.1
pytest-rerunfailures~=11.1.2
pytest-xdist~=3.2.1
pytest-dependency~=0.5.1
pytest-order~=1.1.0
webdriver-manager~=4.0.0
APScheduler~=3.10.1
pytz~=2023.3
python-decouple~=3.8
SQLAlchemy~=2.0.18
pandas~=2.0.0
selenium-wire~=5.1.0
requests~=2.28.2
wcwidth~=0.2.6